[BEAM-1148] Port PAssert away from aggregators
Separates evaluation of the assertion into a transform that outputs
`SuccessOrFailure` from the reporting of failures. The latter happens in
a separate composite transform making it possible to override the
implementation.

Introduces a default implementation that uses Metrics to count the
number of successfully executed assertions as well as the number of
failing assertions.